Created board factory, fixed shooting error
